:fire: added dataclass to Species
