+++ bionty/species/__init__.py
@@ -1,2 +1,2 @@
-from .object import Species
+from .main import Species
 from ._query import update_species_table
